requirements.txt
alembic==1.13.1
asgiref==3.8.1
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.40.0
colorama==0.4.6
Django==5.0.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-jazzmin==3.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mysql-connector-python==8.0.33
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
protobuf==3.20.3
psycopg2==2.9.9
PyJWT==2.8.0
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.30
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
whitenoise==6.7.0
